Update 1.2
Hey hey! Pokemon.java hit 1000 lines! Status effects are now completely implemented and bug free! That's a lie, I didn't implement some of the weird ones and nothing is ever bug free, especially my code. There are now only 4 items left on the to-do list...

+++ src/Trainer.java
@@ -106,7 +106,7 @@ public class Trainer extends GameComponent {
             display += "Slot " + (i + 1) + ": ";
 
             if (this.team[i] != null) {
-                display += this.team[i].getDisplay() + " HP: " + this.team[i].getHp() + "/" + this.team[i].getHpStat();
+                display += this.team[i].getDisplay() + " HP: " + this.team[i].getHp() + "/" + (double)this.team[i].getHpStat();
             } else {
                 display += "Empty";
             }
